# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Estimator9 API",
    "description": "IFRS 9 provisioning engine — MEV series, model fitting, and the accounts that drive them.\n\nObtain a token from `POST /auth/login` and send it as `Authorization: Bearer <token>`. Endpoints marked with a padlock require one; those that also need a permission list it under their 403 response.\n\nTokens are short-lived and carry the caller's permissions, so a role changed after sign-in takes effect on the next token rather than the next request.",
    "contact": {
      "name": "FineIT",
      "email": "akmal@tech.fineit.io"
    },
    "license": {
      "name": ""
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/auth/refresh": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Exchange the refresh cookie for a new access token.",
        "description": "The old refresh token is revoked and replaced, so each one is usable\nexactly once. Presenting an already-used token ends the whole session — see\n[`application::identity::RefreshSession`].\n\nTakes no request body: the credential is the httpOnly cookie, which script\ncannot read and the browser attaches automatically.",
        "operationId": "refresh",
        "responses": {
          "200": {
            "description": "A new session",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing, expired, or already-used refresh token. Reuse ends the whole session",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "The data store is unavailable; retry",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "description": "The body every failing request returns. One shape for all errors, so a\nclient can parse failures without branching on the status code.",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "string",
            "description": "Human-readable description of what went wrong.",
            "example": "user 999999 not found"
          }
        }
      },
...
      "SessionResponse": {
        "type": "object",
        "description": "Returned by both `/auth/login` and `/auth/refresh` — the same thing is\nbeing handed over in each case.",
        "required": [
          "access_token",
          "token_type",
          "expires_at",
          "user"
        ],
        "properties": {
          "access_token": {
            "type": "string",
            "description": "Send as `Authorization: Bearer <access_token>`."
          },
          "expires_at": {
            "type": "string",
            "format": "date-time",
            "description": "When the token stops being accepted. Short by design — permissions are\ncarried inside it, so this bounds how long a revoked role keeps working."
          },
          "token_type": {
            "type": "string",
            "example": "Bearer"
          },
          "user": {
            "$ref": "#/components/schemas/UserResponse",
            "description": "Saves the client an immediate follow-up call to render a header."
          }
        }
      },
...
      "UserResponse": {
        "type": "object",
        "description": "A user as the API publishes them.\n\nThere is deliberately no password field of any kind. Because `User` keeps\nits hash private and this struct is written by hand, adding one would have\nto be a conscious act rather than a `#[derive(Serialize)]` accident.",
        "required": [
          "id",
          "email",
          "name",
          "is_active",
          "roles",
          "is_superadmin",
          "created_at",
          "updated_at"
        ],
        "properties": {
          "created_at": {
            "type": "string",
            "format": "date-time",
            "example": "2026-09-18T12:41:29.785Z"
          },
          "email": {
            "type": "string",
            "example": "analyst@fineit.io"
          },
          "id": {
            "type": "integer",
            "format": "int64",
            "example": 1
          },
          "is_active": {
            "type": "boolean",
            "description": "Deactivated accounts are retained, not deleted.",
            "example": true
          },
          "is_superadmin": {
            "type": "boolean",
            "description": "True when the account holds `superadmin` and therefore bypasses every\npermission check.",
            "example": false
          },
          "name": {
            "type": "string",
            "example": "Test Analyst"
          },
          "roles": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Role names, alphabetically. Permissions are not expanded here — they are\nissued in the access token at login.",
            "example": [
              "analyst"
            ]
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "example": "2026-09-18T12:41:29.785Z"
          }
        }
      }
    },
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Issued by POST /auth/login. Expires; refresh by logging in again."
      }
    }
  },
...
}